Deprecate useTableBulkActions
