--- react/EXPERIMENTAL_useTableBulkActions.ts
@@ -1,3 +1,0 @@
-export {
-  default,
-} from './components/EXPERIMENTAL_Table/hooks/useTableBulkActions'
