+++ react/components/EXPERIMENTAL_useCheckboxTree/constants.ts
@@ -1,0 +1,5 @@
+export const defaultComparatorCurry = (item: any) => (candidate: any) =>
+  item.id && candidate.id && item.id === candidate.id
+
+export const ROOT_KEY = 'VTEX_CheckboxTreeRoot'
+export const ROOT_VALUE = 'ROOT'
